BSA-26 - Rename all Shootaround text to ShootAround and shootaround to shoot-around
