+++ src/redux/reducers/add-shoot-around/add-shoot-around-reducer.ts
