+++ src/redux/reducers/root-reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import applicationReducer from './application/application-reducer';
-import addShootAroundReducer from './add-shootaround/add-shootaround-reducer';
+import addShootAroundReducer from './add-shoot-around/add-shoot-around-reducer';
 import dashboardReducer from './dashboard/dashboard-reducer';
 import shootAroundListReducer from './shoot-around-list/shoot-around-list-reducer';
