+++ src/redux/sagas/root-saga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
 import {applicationMountedWatcherSaga} from './application/application-saga';
-import {submitShootAroundWatcherSaga} from './add-shootaround/add-shootaround-saga';
+import {submitShootAroundWatcherSaga} from './add-shoot-around/add-shoot-around-saga';
 import {dashboardDateRangeChangedWatcherSaga} from './dashboard/dashboard-saga';
 import {getShootAroundListDataWatcherSaga} from './shoot-around-list/shoot-around-list-saga';
