+++ src/screens/shoot-around-list-screen.tsx
@@ -78,7 +78,7 @@ const ShootAroundListScreen = (props: IProps) => {
                 openIcon='close'
                 actions={[
                     {
-                        title: 'Add Shootaround',
+                        title: 'Add Shoot Around',
                         icon: 'basketball-hoop',
                         pressHandler: () => props.navigation.navigate('AddShootAroundScreen')
                     }
